Moved pages out of components subfolder into main folder; moved Page component into core-components, moved routes.js into app root.
