+++ app/core-components/Page/index.js
