+++ app/pages/authentication/SigninPage.js
@@ -4,17 +4,16 @@ import * as React from 'react';
 import { translate } from 'react-i18next';
 import { Grid } from 'semantic-ui-react';
 
+import Page from 'core-components/Page';
 import type { CustomTranslatorProps } from 'types/translator';
 import FlashMessages from 'core-components/flash/FlashMessages';
 import authentication from 'modules/authentication';
 
-import Page from '../../Page';
-
 type Props = CustomTranslatorProps & { /* new props go here */ };
 
-const { ResetCard } = authentication.components;
+const { SigninCard } = authentication.components;
 
-const PureResetPage = (props: Props): React.Node => {
+const PureSigninPage = (props: Props): React.Node => {
   return (
     // $FlowFixMe Can't figure out cause; Page component needs rewriting anyway #TODO
     <Page>
@@ -22,14 +21,14 @@ const PureResetPage = (props: Props): React.Node => {
         <Grid.Column width={6}>
           <FlashMessages />
 
-          <ResetCard />
+          <SigninCard />
         </Grid.Column>
       </Grid>
     </Page>
   );
 };
 
-const ResetPage = translate()(PureResetPage);
+const SigninPage = translate()(PureSigninPage);
 
-export { PureResetPage };
-export default ResetPage;
+export { PureSigninPage };
+export default SigninPage;
