+++ app/pages/authentication/SignupPage.js
@@ -4,17 +4,16 @@ import * as React from 'react';
 import { translate } from 'react-i18next';
 import { Grid } from 'semantic-ui-react';
 
+import Page from 'core-components/Page';
 import type { CustomTranslatorProps } from 'types/translator';
 import FlashMessages from 'core-components/flash/FlashMessages';
 import authentication from 'modules/authentication';
 
-import Page from '../../Page';
-
 type Props = CustomTranslatorProps & { /* new props go here */ };
 
-const { ConfirmCard } = authentication.components;
+const { SignupCard } = authentication.components;
 
-const PureConfirmPage = (props: Props): React.Node => {
+const PureSignupPage = (props: Props): React.Node => {
   return (
     // $FlowFixMe Can't figure out cause; Page component needs rewriting anyway #TODO
     <Page>
@@ -22,14 +21,14 @@ const PureConfirmPage = (props: Props): React.Node => {
         <Grid.Column width={6}>
           <FlashMessages />
 
-          <ConfirmCard />
+          <SignupCard />
         </Grid.Column>
       </Grid>
     </Page>
   );
 };
 
-const ConfirmPage = translate()(PureConfirmPage);
+const SignupPage = translate()(PureSignupPage);
 
-export { PureConfirmPage };
-export default ConfirmPage;
+export { PureSignupPage };
+export default SignupPage;
